vulnerable-apps/api-service/requirements.txt
Flask==2.3.0
psycopg2-binary==2.9.6
requests==2.31.0
PyJWT==2.8.0
